requirements.txt
openai
langchain 
tiktoken 
pypdf 
unstructured 
gradio  
pinecone-client